BaseK2BaseC {far}	R Documentation

Changing Basis

Description

Given the coordinates in the Karhunen-Loève expansion base of the Wiener, compute the coordinates in the canonical basis.

Usage

BaseK2BaseC(x, nb)

Arguments

x
A matrix containing the coordinates in the Karhunen-Loève basis. One observation per column.

nb
The dimension of the canonical basis consider. By default, the dimension is the same as the Karhunen-Loève one (i.e. number of row of x).

Details

The Karhunen-Loève expansion is a sum of an infinity of terms, but here the expansion is truncated to a finite number of terms. Empirically, we remark that using twice the dimension of the canonical basis desired for the number of terms in the expansion is a good compromise.

Value

A object of class fdata with nb discretization points and the same number of observations as x.

Author(s)

J. Damon

References

Pumo, B. (1992). Estimation et Prévision de Processus Autoregressifs Fonctionnels. Applications aux Processus à Temps Continu. PhD Thesis, University Paris 6, Pierre et Marie Curie.

See Also

simul.wiener, simul.far.wiener

Examples

    data1 <- BaseK2BaseC(x=matrix(rnorm(50),ncol=5,nrow=10), nb=5)
    multplot(data1,whole=TRUE)

[Package far version 0.6-7 Index]