base.simul.far {far}	R Documentation

Creating functional basis

Description

Computation of a particular basis in a functional space.

Usage

base.simul.far(m=24, n=5)

Arguments

m
Number of discretization points

n
Number of axis

Details

We consider a sinusoidal basis of the functional space C[0;1] of the continuous functions from [0;1] to R. We compute here the values of the n first (functional) axis at m equi-repartited discretization points in [0;1] (more precisely the point 0,1/\code{m},..., (\code{m}-1)/\code{m}).

Value

A matrix of size m x n containing the m values of the n first axis of the basis.

Note

The chosen basis is orthogonal.

The aim of this function is to provide an internal tool for the function simul.farx.

Author(s)

J. Damon

See Also

simul.farx

Examples

  print(temp<-base.simul.far(10,3))
  print(t(temp)%*%temp)
  matplot(base.simul.far(100,5),type='l')

[Package far version 0.6-7 Index]