date.fdata {far}	R Documentation

Extract the date of fdata

Description

Extract the date(s) of fdata objects

Usage

date.fdata(data)

Arguments

data
A fdata object

Details

The dates are the labels of the functionals observations of the fdata object.

fdata are not constructed as ts object so a specific function to obtain the date is useful.

Value

A vector giving the dates (as character).

Author(s)

J. Damon

See Also

fdata

Examples

# Reading the data
library(stats)
data(UKDriverDeaths)

# Conversion of the data
fUKDriverDeaths <- as.fdata(UKDriverDeaths,col=1,p=12,dates=1969:1984,
                            name="UK Driver Deaths")
date.fdata(fUKDriverDeaths)

[Package far version 0.6-7 Index]