is.na.fdata {far}	R Documentation

Not Available / “Missing” Values

Description

The generic function is.na returns a logical vector of the same “form” as its argument x, containing TRUE for those elements marked NA or NaN (!) and FALSE otherwise. dim, dimnames and names attributes are preserved.

Usage

## S3 method for class 'fdata'
is.na(x)

Arguments

x
A fdata object

Details

An observation is considered as NA if any of its values is NA.

Value

A matrix of Logical values giving as rows the variables of x and as columns the observation.

Author(s)

J. Damon

See Also

NA

Examples

  # Reading of the data
  library(stats)
  data(UKDriverDeaths)
  UKDriverDeaths[20]<-NA

  # Making the data of class 'fdata'
  fUKDriverDeaths <- as.fdata(UKDriverDeaths,col=1,p=12,dates=1969:1984,
                            name="UK Driver Deaths")
  summary(fUKDriverDeaths)
  is.na(fUKDriverDeaths)

[Package far version 0.6-7 Index]