simul.far.sde {far}	R Documentation

FAR-SDE process simulation

Description

Simulation of a FAR process following an Stochastic Differential Equation

Usage

simul.far.sde(coef=c(0.4, 0.8), n=80, p=32, sigma=1)

Arguments

coef
Numerical vertor. It contains the two values of the coefficients (a(1) and a(2), see details for more informations).

n
Integer. The number of observations generated.

p
Integer. The number of discretization points.

sigma
Numeric. The standard deviation (see details for more informations).

Details

This function implements the simulation proposed by Besse and Cardot (1996) to simulate a FAR process following the Stochastic Differential Equation:

d^2(X)+a(2).d(X)+a(1).X=\code{sigma}.d(W)

Where d^2(X) and d(X) stand respectively for the second and first derivate of the process X, and W is a brownian process.

The coefficients a(1) and a(2) are the two first elements of coef.

The simulation use a order one approximation inspired by the work of Milstein, as described in Besse and Cardot (1996).

Value

A fdata object containing one variable ("var") which is a FAR(1) process of length n with p discretization points.

Author(s)

J. Damon

References

Besse, P. and Cardot, H. (1996). Approximation spline de la prévision d'un processus fonctionnel autorégressif d'ordre 1. Revue Canadienne de Statistique/Canadian Journal of Statistics, 24, 467–487.

See Also

simul.far, simul.far.wiener, simul.farx, simul.wiener.

Examples

  far1 <- simul.far.sde()
  summary(far1)
  print(far(far1,kn=2))
  par(mfrow=c(2,1))
  plot(far1,date=1)
  plot(select.fdata(far1,date=1:5),whole=TRUE,separator=TRUE)

[Package far version 0.6-7 Index]