simul.far.wiener {far}	R Documentation

FAR(1) process simulation with Wiener noise

Description

Simulation of a FAR(1) process using a Wiener noise.

Usage

simul.far.wiener(m=64, n=128,
d.rho=diag(c(0.45, 0.9, 0.34, 0.45)), cst1=0.05, m2=NULL)

Arguments

m
Integer. Number of discretization points.

n
Integer. Number of observations.

d.rho
Numerical matrix. Expression of the first bloc of the linear operator in the Karhunen-Loève basis.

cst1
Numeric. Perturbation coefficient on the linear operator.

m2
Integer. Length of the Karhunen-Loève expansion (2m by default).

Details

This function simulate a FAR(1) process with a Wiener noise. As for the simul.wiener, the function use the Karhunen-Loève expansion of the noise. The FAR(1) process, defined by its linear operator (see far for more details), is computed in the Karhunen-Loève basis then projected in the natural basis. The parameters given in input (d.rho and cst1) are expressed in the Karhunen-Loève basis.

The linear operator, expressed in the Karhunen-Loève basis, is of the form:

\code{d.rho} 0 \cr 0 eps.rho

Where d.rho is the matrix provided in ths call, the two 0 are in fact two blocks of 0, and eps.rho is a diagonal matrix having on his diagonal the terms:

(eps(k+1), eps(k+2), …, eps(\code{m2}))

where

eps(i)= \code{cst1}/(i^2)+(1-\code{cst1})/exp(i)

and k is the length of the d.rho diagonal.

The d.rho matrix can be viewed as the information and the eps.rho matrix as a perturbation. In this logic, the norm of eps.rho need to be smaller than the one of d.rho.

Value

A fdata object containing one variable ("var") which is a FAR(1) process of length n with m discretization points.

Author(s)

J. Damon

References

Pumo, B. (1992). Estimation et Prévision de Processus Autoregressifs Fonctionnels. Applications aux Processus à Temps Continu. PhD Thesis, University Paris 6, Pierre et Marie Curie.

See Also

fdata, far , simul.far.wiener.

Examples

  far1 <- simul.far.wiener(m=64,n=100)
  summary(far1)
  print(far(far1,kn=4))
  par(mfrow=c(2,1))
  plot(far1,date=1)
  plot(select.fdata(far1,date=1:5),whole=TRUE,separator=TRUE)

[Package far version 0.6-7 Index]