simul.wiener {far}	R Documentation

Wiener process simulation

Description

Simulation of Wiener processes.

Usage

simul.wiener(m=64, n=1, m2=NULL)

Arguments

m
Integer. Number of discretization points.

n
Integer. Number of observations.

m2
Integer. Length of the Karhunen-Loève expansion (2m by default).

Details

This function use the known Karhunen-Loève expansion of Wiener processes to simulate observations of such a process.

The option m2 is internally used to set the length of the expansion. This expansion need to be larger than the number of discretization points, but a too important value may slow down the generation. The default value as been chosen as a compromise.

Value

A fdata object containing one variable ("var") which is a Wiener process of length n with m discretization points.

Author(s)

J. Damon

References

Pumo, B. (1992). Estimation et Prévision de Processus Autoregressifs Fonctionnels. Applications aux Processus à Temps Continu. PhD Thesis, University Paris 6, Pierre et Marie Curie.

See Also

simul.far.sde, simul.far.wiener, simul.farx, simul.far.

Examples

  noise <- simul.wiener(m=64,n=100,m2=512)
  summary(noise)
  par(mfrow=c(2,1))
  plot(noise,date=1)
  plot(select.fdata(noise,date=1:5),whole=TRUE,separator=TRUE)

[Package far version 0.6-7 Index]